Collect post code changes, change type to Long

+++ app/src/main/java/com/papageorgiouk/curfnsurf/ui/form/postcode/PostCodeViewModel.kt
@@ -5,7 +5,7 @@ import com.papageorgiouk.curfnsurf.data.FormManager
 
 class PostCodeViewModel(private val formManager: FormManager) : ViewModel() {
 
-    fun setPostCode(postCode: Int) {
+    fun setPostCode(postCode: Long) {
         formManager.postCode = postCode
     }
